@@ -1,6 +1,6 @@
 {
   "name": "pg-protocol",
-  "version": "1.2.5",
+  "version": "1.5.0",
   "description": "The postgres client/server binary protocol, implemented in TypeScript",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -13,7 +13,7 @@
     "chunky": "^0.0.0",
     "mocha": "^7.1.2",
     "ts-node": "^8.5.4",
-    "typescript": "^3.7.3"
+    "typescript": "^4.0.3"
   },
   "scripts": {
     "test": "mocha dist/**/*.test.js",
@@ -21,5 +21,14 @@
     "build:watch": "tsc --watch",
     "prepublish": "yarn build",
     "pretest": "yarn build"
-  }
+  },
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/brianc/node-postgres.git",
+    "directory": "packages/pg-protocol"
+  },
+  "files": [
+    "/dist/*{js,ts,map}",
+    "/src"
+  ]
 }
