@@ -1,17 +1,18 @@
-var pg = require('pg')
-var assert = require('assert')
-var QueryStream = require('../')
+import pg from 'pg'
+import assert from 'assert'
+import QueryStream from '../src'
 
 describe('client options', function () {
   it('uses custom types from client config', function (done) {
     const types = {
       getTypeParser: () => (string) => string,
     }
-    var client = new pg.Client({ types })
+    //@ts-expect-error
+    const client = new pg.Client({ types })
     client.connect()
-    var stream = new QueryStream('SELECT * FROM generate_series(0, 10) num')
-    var query = client.query(stream)
-    var result = []
+    const stream = new QueryStream('SELECT * FROM generate_series(0, 10) num')
+    const query = client.query(stream)
+    const result = []
     query.on('data', (datum) => {
       result.push(datum)
     })
