@@ -1,16 +1,18 @@
-var assert = require('assert')
-var concat = require('concat-stream')
-
-var QueryStream = require('../')
-var helper = require('./helper')
+import assert from 'assert'
+import concat from 'concat-stream'
+import QueryStream from '../src'
+import helper from './helper'
 
 if (process.version.startsWith('v8.')) {
   console.error('warning! node less than 10lts stream closing semantics may not behave properly')
 } else {
   helper('close', function (client) {
     it('emits close', function (done) {
-      var stream = new QueryStream('SELECT * FROM generate_series(0, $1) num', [3], { batchSize: 2, highWaterMark: 2 })
-      var query = client.query(stream)
+      const stream = new QueryStream('SELECT * FROM generate_series(0, $1) num', [3], {
+        batchSize: 2,
+        highWaterMark: 2,
+      })
+      const query = client.query(stream)
       query.pipe(concat(function () {}))
       query.on('close', done)
     })
@@ -18,12 +20,12 @@ if (process.version.startsWith('v8.')) {
 
   helper('early close', function (client) {
     it('can be closed early', function (done) {
-      var stream = new QueryStream('SELECT * FROM generate_series(0, $1) num', [20000], {
+      const stream = new QueryStream('SELECT * FROM generate_series(0, $1) num', [20000], {
         batchSize: 2,
         highWaterMark: 2,
       })
-      var query = client.query(stream)
-      var readCount = 0
+      const query = client.query(stream)
+      let readCount = 0
       query.on('readable', function () {
         readCount++
         query.read()
@@ -38,7 +40,7 @@ if (process.version.startsWith('v8.')) {
     })
 
     it('can destroy stream while reading', function (done) {
-      var stream = new QueryStream('SELECT * FROM generate_series(0, 100), pg_sleep(1)')
+      const stream = new QueryStream('SELECT * FROM generate_series(0, 100), pg_sleep(1)')
       client.query(stream)
       stream.on('data', () => done(new Error('stream should not have returned rows')))
       setTimeout(() => {
@@ -48,7 +50,7 @@ if (process.version.startsWith('v8.')) {
     })
 
     it('emits an error when calling destroy with an error', function (done) {
-      var stream = new QueryStream('SELECT * FROM generate_series(0, 100), pg_sleep(1)')
+      const stream = new QueryStream('SELECT * FROM generate_series(0, 100), pg_sleep(1)')
       client.query(stream)
       stream.on('data', () => done(new Error('stream should not have returned rows')))
       setTimeout(() => {
@@ -63,7 +65,7 @@ if (process.version.startsWith('v8.')) {
     })
 
     it('can destroy stream while reading an error', function (done) {
-      var stream = new QueryStream('SELECT * from  pg_sleep(1), basdfasdf;')
+      const stream = new QueryStream('SELECT * from  pg_sleep(1), basdfasdf;')
       client.query(stream)
       stream.on('data', () => done(new Error('stream should not have returned rows')))
       stream.once('error', () => {
@@ -74,7 +76,7 @@ if (process.version.startsWith('v8.')) {
     })
 
     it('does not crash when destroying the stream immediately after calling read', function (done) {
-      var stream = new QueryStream('SELECT * from generate_series(0, 100), pg_sleep(1);')
+      const stream = new QueryStream('SELECT * from generate_series(0, 100), pg_sleep(1);')
       client.query(stream)
       stream.on('data', () => done(new Error('stream should not have returned rows')))
       stream.destroy()
@@ -82,7 +84,7 @@ if (process.version.startsWith('v8.')) {
     })
 
     it('does not crash when destroying the stream before its submitted', function (done) {
-      var stream = new QueryStream('SELECT * from generate_series(0, 100), pg_sleep(1);')
+      const stream = new QueryStream('SELECT * from generate_series(0, 100), pg_sleep(1);')
       stream.on('data', () => done(new Error('stream should not have returned rows')))
       stream.destroy()
       stream.on('close', done)
