@@ -7,10 +7,9 @@ const it = require('mocha').it
 const Pool = require('../')
 
 describe('verify', () => {
-  it('verifies a client with a callback', false, (done) => {
+  it('verifies a client with a callback', (done) => {
     const pool = new Pool({
       verify: (client, cb) => {
-        client.release()
         cb(new Error('nope'))
       },
     })
