@@ -1,3 +1,4 @@
+const assert = require('assert')
 const helper = require('../test-helper')
 const suite = new helper.Suite()
 const { Client } = helper.pg
@@ -8,6 +9,7 @@ suite.test('it sends options', async () => {
   })
   await client.connect()
   const { rows } = await client.query('SHOW default_transaction_isolation')
-  console.log(rows)
+  assert.strictEqual(rows.length, 1)
+  assert.strictEqual(rows[0].default_transaction_isolation, 'serializable')
   await client.end()
 })
