@@ -5,6 +5,7 @@ var util = require('util')
 
 var pg = helper.pg
 const Client = pg.Client
+const DatabaseError = pg.DatabaseError
 
 var createErorrClient = function () {
   var client = helper.client()
@@ -19,7 +20,10 @@ const suite = new helper.Suite('error handling')
 
 suite.test('sending non-array argument as values causes an error callback', (done) => {
   const client = new Client()
-  client.connect(() => {
+  client.connect((err) => {
+    if (err) {
+      return done(err)
+    }
     client.query('select $1::text as name', 'foo', (err) => {
       assert(err instanceof Error)
       client.query('SELECT $1::text as name', ['foo'], (err, res) => {
@@ -32,7 +36,10 @@ suite.test('sending non-array argument as values causes an error callback', (don
 
 suite.test('re-using connections results in error callback', (done) => {
   const client = new Client()
-  client.connect(() => {
+  client.connect((err) => {
+    if (err) {
+      return done(err)
+    }
     client.connect((err) => {
       assert(err instanceof Error)
       client.end(done)
@@ -40,19 +47,22 @@ suite.test('re-using connections results in error callback', (done) => {
   })
 })
 
-suite.test('re-using connections results in promise rejection', (done) => {
+suite.testAsync('re-using connections results in promise rejection', () => {
   const client = new Client()
-  client.connect().then(() => {
-    client.connect().catch((err) => {
+  return client.connect().then(() => {
+    return helper.rejection(client.connect()).then((err) => {
       assert(err instanceof Error)
-      client.end().then(done)
+      return client.end()
     })
   })
 })
 
 suite.test('using a client after closing it results in error', (done) => {
   const client = new Client()
-  client.connect(() => {
+  client.connect((err) => {
+    if (err) {
+      return done(err)
+    }
     client.end(
       assert.calls(() => {
         client.query(
@@ -131,6 +141,9 @@ suite.test('when a query is binding', function (done) {
   )
 
   assert.emits(query, 'error', function (err) {
+    if (!helper.config.native) {
+      assert(err instanceof DatabaseError)
+    }
     assert.equal(err.severity, 'ERROR')
     ensureFuture(client, done)
   })
@@ -204,6 +217,9 @@ suite.test('within a simple query', (done) => {
   var query = client.query(new pg.Query("select eeeee from yodas_dsflsd where pixistix = 'zoiks!!!'"))
 
   assert.emits(query, 'error', function (error) {
+    if (!helper.config.native) {
+      assert(error instanceof DatabaseError)
+    }
     assert.equal(error.severity, 'ERROR')
     done()
   })
@@ -227,12 +243,16 @@ suite.test('connected, idle client error', (done) => {
 
 suite.test('cannot pass non-string values to query as text', (done) => {
   const client = new Client()
-  client.connect()
-  client.query({ text: {} }, (err) => {
-    assert(err)
-    client.query({}, (err) => {
-      client.on('drain', () => {
-        client.end(done)
+  client.connect((err) => {
+    if (err) {
+      return done(err)
+    }
+    client.query({ text: {} }, (err) => {
+      assert(err)
+      client.query({}, (err) => {
+        client.on('drain', () => {
+          client.end(done)
+        })
       })
     })
   })
