@@ -32,7 +32,7 @@ let makeTerminatingBackend = (byte) => {
 
 suite.test('SSL connection error allows event loop to exit', (done) => {
   const port = makeTerminatingBackend('N')
-  const client = new helper.pg.Client({ ssl: 'require', port })
+  const client = new helper.pg.Client({ ssl: 'require', port, host: 'localhost' })
   // since there was a connection error the client's socket should be closed
   // and the event loop will have no refs and exit cleanly
   client.connect((err) => {
@@ -43,7 +43,7 @@ suite.test('SSL connection error allows event loop to exit', (done) => {
 
 suite.test('Non "S" response code allows event loop to exit', (done) => {
   const port = makeTerminatingBackend('X')
-  const client = new helper.pg.Client({ ssl: 'require', port })
+  const client = new helper.pg.Client({ ssl: 'require', host: 'localhost', port })
   // since there was a connection error the client's socket should be closed
   // and the event loop will have no refs and exit cleanly
   client.connect((err) => {
